Read/Write functions work properly now.
The read and write functions are now finished. The functions
now can write and read data, no matter the amount of bytes to
be written/readed or if more than one page is needed to do this.
The functions are able to change the page to the next one if
they have not finished the write/read task.

Both functions were tested with 4 test cases, successfully.
But this version still has print messages for debugging.